fix some merge conflicts in pantries backend
