+++ apps/backend/src/users/user.entity.ts
@@ -6,7 +6,7 @@ import { Role } from './types';
 export class User {
   @PrimaryGeneratedColumn({ name: 'user_id' })
   id: number;
-  
+
   @Column({
     type: 'varchar',
     length: 20,
@@ -28,8 +28,4 @@ export class User {
     length: 20,
   })
   phone: string;
-
-  get id(): number {
-    return this.user_id;
-  }
 }
